requirements.txt
boto3==1.9.*
slackclient==2.0.1
